backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
openai==1.50.0
httpx==0.27.0
Pillow==10.4.0
python-dotenv==1.0.1
rembg==2.0.60
numpy==1.26.4
imageio-ffmpeg==0.4.9
